+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix `Event::to_static` returning `None` for user events.
 - On Wayland, Hide CSD for fullscreen windows.
 - On Windows, ignore spurious mouse move messages.
+- **Breaking:** Move `ModifiersChanged` variant from `DeviceEvent` to `WindowEvent`.
 
 # 0.21.0 (2020-02-04)